% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elsocr_load.R
\name{elsocr_load}
\alias{elsocr_load}
\title{Cargar base de datos ELSOC al entorno global}
\usage{
elsocr_load()
}
\value{
sin valor asociado
}
\description{
Cargar base de datos ELSOC al entorno global
}
\examples{
nombre_objeto <- elsocr_load()
}
